system-plugin: move API code to plugin/api directory
