+++ src/plugin/api/system/authentication/store.c
