+++ src/plugin/api/system/authentication/store.h
