+++ src/plugin/api/system/store.c
