+++ src/plugin/startup/store.c
@@ -19,15 +19,15 @@
 #include "srpc/common.h"
 #include "srpc/ly_tree.h"
 #include "srpc/types.h"
-#include "system/api/authentication/check.h"
-#include "system/api/authentication/load.h"
-#include "system/api/check.h"
-#include "system/api/ntp/check.h"
-#include "system/api/ntp/store.h"
-#include "system/api/store.h"
-#include "system/api/dns_resolver/store.h"
-#include "system/api/dns_resolver/check.h"
-#include "system/api/authentication/store.h"
+#include "plugin/api/system/authentication/check.h"
+#include "plugin/api/system/authentication/load.h"
+#include "plugin/api/system/check.h"
+#include "plugin/api/system/ntp/check.h"
+#include "plugin/api/system/ntp/store.h"
+#include "plugin/api/system/store.h"
+#include "plugin/api/system/dns_resolver/store.h"
+#include "plugin/api/system/dns_resolver/check.h"
+#include "plugin/api/system/authentication/store.h"
 
 // data manipulation
 #include "system/data/authentication/authorized_key.h"
